+++ jeap-deploymentlog-docgen/src/main/java/ch/admin/bit/jeap/deploymentlog/docgen/ConfluenceAdapterMock.java
@@ -13,14 +13,6 @@ public class ConfluenceAdapterMock implements ConfluenceAdapter {
     private final Set<String> modifiedPages = new ConcurrentSkipListSet<>();
 
     @Override
-    public String getPageByName(String pageName) {
-        int fakePageId = pageName.hashCode();
-        log.info("Get page by name: pageName={} fakePageId={}", pageName, fakePageId);
-        modifiedPages.add(pageName);
-        return String.valueOf(fakePageId);
-    }
-
-    @Override
     public String addOrUpdatePageUnderAncestor(String ancestorId, String pageName, String content) {
         int fakePageId = (ancestorId + pageName).hashCode();
         log.info("Add or update page: ancestorId={} pageName={} fakePageId={}", ancestorId, pageName, fakePageId);